requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
aiofiles
sqlalchemy<2.0
databases
databases[aiosqlite]

python-jose[cryptography]
passlib[bcrypt]
pytest
pytest-asyncio
httpx